Update and rename _item_sets_pvp.go to item_sets_pvp.go
