+++ sim/paladin/item_sets_pvp.go
@@ -1,0 +1,46 @@
+package paladin
+
+import (
+	"github.com/wowsims/classic/sim/core"
+	"github.com/wowsims/classic/sim/core/stats"
+)
+
+var KnightLieutenantsLamellarPlate = core.NewItemSet(core.ItemSet{
+	Name: "Lieutenant Commander's Redoubt",
+	Bonuses: map[int32]core.ApplyEffect{
+		// Increases damage and healing done by magical spells and effects by up to 23.
+		2: func(agent core.Agent) {
+			c := agent.GetCharacter()
+			c.AddStat(stats.SpellPower, 23)
+		},
+		// Reduces the cooldown of your Hammer of Justice by 10 sec.
+		4: func(agent core.Agent) {
+			// Nothing to do
+		},
+		// +20 Stamina.
+		6: func(agent core.Agent) {
+			c := agent.GetCharacter()
+			c.AddStat(stats.Stamina, 20)
+		},
+	},
+})
+
+var KnightLieutenantsLamellarPlate = core.NewItemSet(core.ItemSet{
+	Name: "Field Marshal's Aegis",
+	Bonuses: map[int32]core.ApplyEffect{
+		// +20 Stamina.
+		2: func(agent core.Agent) {
+			c := agent.GetCharacter()
+			c.AddStat(stats.Stamina, 20)
+		},
+		// Reduces the cooldown of your Hammer of Justice by 10 sec.
+		4: func(agent core.Agent) {
+			// Nothing to do
+		},
+		// Increases damage and healing done by magical spells and effects by up to 23.
+		6: func(agent core.Agent) {
+			c := agent.GetCharacter()
+			c.AddStat(stats.SpellPower, 23)
+		},
+	},
+})
